Fix some function in js
